src/cf-ingest-db/requirements.txt
functions-framework==3.7.0
pandas==2.2.1
psycopg2-binary==2.9.9
boto3==1.34.22
numpy==1.26.4
pyarrow==15.0.2
google-cloud-secret-manager==2.26.0
python-dotenv==1.0.1
